requirements.txt
flask
flask-restx
pathlib
python-dotenv
flask-jwt-extended
flask-sqlalchemy
flask-Migrate
flask-cors
mysqlclient
psycopg2-binary